chore: merge PlayerContainer to Player
