+++ example/package.json
@@ -10,7 +10,7 @@
     "griffith": "^1.17.0",
     "react": "^16.8",
     "react-dom": "^16.8",
-    "react-router-dom": "^5.2.0"
+    "react-router-dom": "^5.3.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react-refresh": "^1.3.3",
