+++ packages/griffith/src/index.ts
@@ -1,5 +1,5 @@
-export {default} from './components/PlayerContainer'
-export * from './components/PlayerContainer'
+export {default} from './components/Player'
+export * from './components/Player'
 export {ExternalContext as MessageContext} from './contexts/MessageContext'
 export {default as VideoSourceContext} from './contexts/VideoSourceContext'
 export {default as Controller} from './components/Controller'
