+++ tsconfig.json
@@ -15,6 +15,8 @@
     "declaration": true,
     // import @types/react
     "allowSyntheticDefaultImports": true,
+    // import aphrodite/no-important
+    "moduleResolution": "Node",
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true
